djangotutorial/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
pandas==2.2.3
plotly==5.24.1
Requests==2.32.3
